+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [0.4.3] - 2025-06-08
+### Added
+- Docs: installation from compose
+- Custom access & blocked log
+### Change
+- Default console format
+
+## [0.4.2] - 2025-06-06
+### Added
+- Custom 403 page for unauthorized IPs
+### Change
+- Rework code format (server.py, handlers/)
+
 ## [0.4.1] - 2025-06-05
 ### Added
 - Custom log format in config.ini